feature: added GeneratedConsumerTokenAuditDetails data model and audit.ts file utils (PIN-10577)
