+++ packages/authorization-server/.env
@@ -27,7 +27,8 @@ RATE_LIMITER_REDIS_HOST="localhost"
 RATE_LIMITER_REDIS_PORT="6379"
 RATE_LIMITER_TIMEOUT_MILLIS="300"
 
-S3_BUCKET=interop-local-bucket
+S3_BUCKET_CONSUMER_TOKEN_AUDIT_FALLBACK=interop-local-bucket
+S3_BUCKET_API_TOKEN_AUDIT_FALLBACK=interop-local-bucket
 S3_CUSTOM_SERVER=true
 S3_SERVER_HOST=http://localhost
 S3_SERVER_PORT=9000
