+++ packages/token-details-persister/test/consumerService.integration.test.ts
@@ -2,7 +2,7 @@
 import crypto from "crypto";
 import {
   algorithm,
-  GeneratedTokenAuditDetails,
+  GeneratedConsumerTokenAuditDetails,
   generateId,
 } from "pagopa-interop-models";
 import { afterAll, beforeAll, describe, expect, it, vi } from "vitest";
@@ -72,7 +72,7 @@ describe("consumerService", () => {
   });
 
   it("should write three entries on the bucket", async () => {
-    const auditMessages: GeneratedTokenAuditDetails[] = [
+    const auditMessages: GeneratedConsumerTokenAuditDetails[] = [
       getMockAuditDetails(),
       getMockAuditDetails(),
       getMockAuditDetails(),
@@ -150,7 +150,7 @@ describe("consumerService", () => {
   });
 });
 
-const getMockAuditDetails = (): GeneratedTokenAuditDetails => ({
+const getMockAuditDetails = (): GeneratedConsumerTokenAuditDetails => ({
   correlationId: generateId(),
   eserviceId: generateId(),
   descriptorId: generateId(),
